django/requirements.txt
anyio==3.3.0
arrow==1.1.1
asgiref==3.4.1
attrs==21.2.0
binaryornot==0.4.4
black==21.9b0
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.4
click==8.0.1
cookiecutter==1.7.3
coverage==5.5
cryptography==3.4.7
defusedxml==0.7.1
dj-rest-auth==2.1.11
Django==3.2.8
django-allauth==0.45.0
django-cors-headers==3.8.0
django-htmx-rest==0.0.1b2
django-tailwind==2.2.0
django-types==0.10.0
djangorestframework==3.12.4
google-api-core==2.0.0
google-api-python-client==2.32.0
google-api-python-client-stubs==1.9.0
google-auth==2.0.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.5
googleapis-common-protos==1.53.0
gunicorn==20.1.0
h11==0.12.0
httpcore==0.13.6
httplib2==0.19.1
httpx==0.19.0
idna==3.2
iniconfig==1.1.1
Jinja2==3.0.1
jinja2-time==0.2.0
MarkupSafe==2.0.1
mypy-extensions==0.4.3
oauthlib==3.1.1
packaging==21.2
pathspec==0.9.0
platformdirs==2.4.0
pluggy==1.0.0
poyo==0.5.0
protobuf==3.17.3
psycopg2-binary==2.9.2
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==2.1.0
pyparsing==2.4.7
pytest==6.2.5
pytest-django==4.4.0
python-dateutil==2.8.2
python-slugify==5.0.2
python3-openid==3.2.0
pytz==2021.1
regex==2021.10.23
requests==2.26.0
requests-oauthlib==1.3.0
rfc3986==1.5.0
rsa==4.7.2
six==1.16.0
sniffio==1.2.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
tomli==1.2.2
types-pytz==2021.3.1
types-PyYAML==6.0.1
typing-extensions==3.10.0.2
uritemplate==3.0.1
urllib3==1.26.6
